requirements.txt
pdfminer.six
PyMuPDF
